// src/app/core/interfaces/aminoacido.interface.ts
export interface Aminoacido {
  id?: number;
  acidoAspartico?: number;
  treonina?: number;
  serina?: number;
  acidoGlutaminico?: number;
  prolina?: number;
  glicina?: number;
  alanina?: number;
  cisteina?: number;
  valina?: number;
  metionina?: number;
  isoleucina?: number;
  leucina?: number;
  tirosina?: number;
  fenilalanina?: number;
  histidina?: number;
  lisina?: number;
  arginina?: number;
  triptofano?: number;
}